Unit testing on @Suhan and @Abigails functions
Unit testing: IN PROGRESS
@Suhan
The meal functions are primarily working. The issue we have is that
there is
an incorrect number of inquiries from the program. The program only asks
for first, last, and common days. We want the program to ask for every
meal separately. We can easily do this by creating a for loop for
i>numOfDays; then an additional for loop for j < 3; for breakfast, lunch
and dinner. You are also printing the Allow and Extra, and they both
print 0. You can send main the total meal cost for all the meals added
up. Or i see that you have done some tallyicorporate this, that's fine
too. However, I will ang up for the allowed and extra. If you want to
inlready be calculating totals in the main to reduce redundancy; we can
have the total meal cost sent back to the main. You are also asking for
the departure and arrival time twice.
@Abigail
You just hopped on the teams to discuss what we can do with your
section.